chore: refactor interpreter to use an environment for variable storage
The interpreter has been refactored to use an `Environment` struct for variable storage. This allows for defining and assigning variables within the interpreter. The `Environment` struct has been added to the interpreter module and is used to store variable values. The `Interpreter` struct now contains an instance of the `Environment` struct. The `interpret_stmt` method has been added to the `Interpreter` struct to interpret a list of statements. The `execute` method has been added to the `Interpreter` struct to execute individual statements. The `Var` statement has been added to the `Stmt` enum to represent variable declarations. The `evaluate_expr` method has been renamed from `evaluate` to better reflect its purpose. The `evaluate_expr` method now returns a `Result` type instead of panicking on error. The `evaluate_expr` method has been updated to handle the `Assign` expression, allowing for variable assignment.